Upated all dependencies | Twitter login working | v1.62c

+++ MosquitoAlert/tigabib/src/main/java/ceab/movelab/tigabib/ContProvValuesReports.java
@@ -69,16 +69,16 @@ public class ContProvValuesReports {
 		initialValues.put(Reports.KEY_LOCATION_CHOICE, rep.locationChoice);
 
 		if (rep.currentLocationLat != null)
-			initialValues.put(Reports.KEY_CURRENT_LOCATION_LAT, rep.currentLocationLat.floatValue());
+			initialValues.put(Reports.KEY_CURRENT_LOCATION_LAT, rep.currentLocationLat);
 
 		if (rep.currentLocationLon != null)
-			initialValues.put(Reports.KEY_CURRENT_LOCATION_LON, rep.currentLocationLon.floatValue());
+			initialValues.put(Reports.KEY_CURRENT_LOCATION_LON, rep.currentLocationLon);
 
 		if (rep.selectedLocationLat != null)
-			initialValues.put(Reports.KEY_SELECTED_LOCATION_LAT, rep.selectedLocationLat.floatValue());
+			initialValues.put(Reports.KEY_SELECTED_LOCATION_LAT, rep.selectedLocationLat);
 
 		if (rep.selectedLocationLon != null)
-			initialValues.put(Reports.KEY_SELECTED_LOCATION_LON, rep.selectedLocationLon.floatValue());
+			initialValues.put(Reports.KEY_SELECTED_LOCATION_LON, rep.selectedLocationLon);
 
 		initialValues.put(Reports.KEY_PHOTO_ATTACHED, rep.photoAttached);
